+++ src/main/java/br/pucrio/les/esg_token_backend/resources/auth/services/TokenIssuerService.java
@@ -8,6 +8,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import br.pucrio.les.esg_token_backend.resources.users.model.User;
+import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 
@@ -35,6 +36,20 @@ public class TokenIssuerService {
 
         return Jwts.builder().setIssuer(this.issuer).setSubject(usuario.getId().toString()).setIssuedAt(new Date())
                 .setExpiration(exp).signWith(SignatureAlgorithm.HS256, secret).compact();
+    }
+
+    public boolean isTokenValid(String token) {
+        try {
+            Jwts.parser().setSigningKey(secret).parseClaimsJws(token);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    public Long getTokenId(String token) {
+        Claims body = Jwts.parser().setSigningKey(secret).parseClaimsJws(token).getBody();
+        return Long.valueOf(body.getSubject());
     }
 
 }